06-async-deep/pyproject.toml
[project]
name = "async-deep"
version = "0.1.0"
description = "06 단계 — asyncio 심화 (이벤트 루프 / gather / TaskGroup / sync vs async / 부하 비교)"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",        # async I/O 데모용
    "fastapi-hello",      # t07 부하 비교 — 04 의 앱 인메모리에서 호출
]

[tool.uv.sources]
fastapi-hello = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/asyncdeep"]
